renommage dao, ajout des longs, double version pagination

+++ src/com/excilys/database/DAO.java
@@ -8,11 +8,11 @@ public interface DAO<T> {
 
 	public List<T> getAll() throws SQLException;
 	
-	public T get(int id) throws SQLException;
+	public T get(long id) throws SQLException;
 	
 	public void create(T c) throws SQLException;
 	
 	public void update(T c) throws SQLException;
 	
-	public void delete(int c) throws SQLException;
+	public void delete(long c) throws SQLException;
 }
